+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GoogleStorageInterpreter.scala
@@ -20,7 +20,7 @@ import com.google.cloud.storage.Storage.{
   BucketGetOption,
   BucketSourceOption
 }
-import com.google.cloud.storage.{Acl, Blob, BlobId, BlobInfo, Bucket, BucketInfo, Storage, StorageOptions}
+import com.google.cloud.storage.{Acl, Blob, BlobId, BlobInfo, BucketInfo, Storage, StorageOptions}
 import com.google.cloud.{Identity, Policy, Role}
 import fs2.{text, Pipe, Stream}
 import org.typelevel.log4cats.StructuredLogger
@@ -383,9 +383,9 @@ private[google2] class GoogleStorageInterpreter[F[_]](
                          bucketName: GcsBucketName,
                          bucketGetOptions: List[BucketGetOption] = List.empty,
                          traceId: Option[TraceId] = None
-  ): F[Option[Bucket]] = {
+  ): F[Option[BucketInfo]] = {
     val dbForProject = db.getOptions.toBuilder.setProjectId(googleProject.value).build().getService
-    val fa = Async[F].delay(dbForProject.get(bucketName.value, bucketGetOptions: _*)).map(Option(_))
+    val fa: F[Option[BucketInfo]] = Async[F].delay(dbForProject.get(bucketName.value, bucketGetOptions: _*)).map(Option(_))
     withLogging(fa, traceId, s"com.google.cloud.storage.Storage.get(${bucketName.value}, $bucketGetOptions)")
   }